[IN-PROGRESS] Interpret variable statements
Bit the bullet and copied the string to the scanner, so we
got rid of some lifetimes. I tried to use unsafe and gave up
and just decided to use a Vec. There's no clean way to point
at an array of chars underlying a string.

Next steps will be to get objects working