Create GraphViz debug printer
The lox implementation from the book prints bytecode as a
table, which I've create a method to do; I just haven't
written the content of that method.

This implementation though also includes a graphviz output
so that you can view the content of the bytecode as a graph.

This will be especially useful when there are jump ops, because
they can point back to thenselves.

A similar output of the memory will also be useful, though
it will need to use the garbage collector to really be
useful.