Quick add README to document graphviz
The GraphViz functionality in the code is targeted towards
developers, so it doesn't need to be documented in the
code itself. Rather, a note in the README is enough to
let people know how to use it for debugging.

This is just a really quick note to myself in the README,
I'll fill it out more later.