[IN-PROGRESS] Begin to use references for functions
A few things right off the bat that won't stick
around after this change:

1. Lack of a closure type is going to require pretty big changes
2. Everything is split between mut and non mut
3. The thing doesn't run because there's no run method

I think that pulling objects from memory is great, but probably
I'll need to start using Arc's if I really want to pass things
the way I've been doing. I don't know that it's going to be
possible to use `&` or `&mut` for everything.

As for the positive statement of what I *did* do:

1. Add a test that just interprets a basic program
2. Make changes until it compiles, including all of the following
3. Use refs instead of functions so I can swap out the function easier
4. Start to use call frames, though I don't think what I've got here feels quite right