[IN-PROGRESS] Begin executing functions
This is a relatively simple change. It still doesn't work even
though it builds. It's waiting on the implementations of Into
that I made for u8 and Op for one another in another branch.

That branch will be the next commit because I'm about to rebase
it over this.

There's not much else to be said, though this is the first
commit that contains the main loop of the VM, yay!