Visualize OpAdd and OpNegate
This is a really small change, just a step on the way to getting
the first fully running example. It basically re-maps the existing
Ops. The error now is to do with OpConstant, which I'll implement.

I also want to add in String interning pretty soon... so keep
an eye out for that.