Use Box in object type
This way all the objects will contain pointers instead of
the full data, including Strings which could be arbitrarily
big. I'm not sure the internal layout of String or how that
could affect size, but I do know that some of these structs
will be arbitrarily big, so I'm going to just use a pointer
and not worry about size.