String interning with HashMap
I did this entirely without looking at the reference implementation,
and built entirely off of the Rust standard hash map. That said,
I think most of it will stay in place. The only thing that may be
tricky here is figuring out which of the interned strings should be
released in the garbage collector. This could especially become a
problem if there are a large number of strings. I can't remember
how the book resolves this.

Maybe interned strings are released whenever the underlying memory
value is released...