Completed running VM for some operations
Finally working: I've got the full runtime functionality done
for some features. Now I can add things into the compiler and
the VM separately with tests, but with only small changes
to existing methods for the most part. (Closures and Upvalues
will still be relatively big changes, as will GC).

Things that are working now:

- Constant numbers
- Constant strings
- String interning
- Printing numbers and strings

Also in this change: I put the code/constants into the chunk
rather than into the function. I'm not sure if there's any real
value in this---maybe when we re-compile functions---but I think
it'll become necessary and is easier now. It also made some of
the function manipulation where I get the constants out easier,
and match with the C version.