Implement assert operation
This is required in order to build more sophisticated tests
that actually test the result of lox operations rather than
just the ability to compile and run without error.

The gist is that we add an assert operation which exits with
a failure code if the argument is either false or nil, otherwise
it will just pop the argument.

This change also updates error handling and line tracking so
that assertion printing works more like expected.