Begin to implement binary operations
No unit tests in this commit because I was focused on getting the
graphing bit working (partially, the graph still needs to point
to actual constant values rather than the pointers). This commit
just adds the operations, unit tests will come later to verify
the bytecode produced and behavior of that byteode.

Also, this is pretty important and I will get around to it
eventually I promise, I still need to get the prompt to run the
last compiled function rather than always running the first line