Fix line numbers and running new functions
The Repl only ever ran the first function, and that's because I
had always pointed it to memory location `0`. This just points to
the memory of the last compiled funciton.

Also, the lines in the error messages were sometimes off by
one, because they pointed to the current instead of the previous
token. Now, this means that the last token in an expression is the
one whose line number is reported, but that will at least give
the user a rough idea and will always work for one-line expresisons