Compile equality comparison and addition
The primary purpose of this change is to add and fix two tests
that compile addition and comparison. They just compile and
check the bytecode; they don't actually run the programs. Running
them would result in errors, so that will go in the next commit.

This change also renames all of the tests to make it easier to
filter them without including `test_` in the name.