Implement comparison in interpreter
Comparison for booleans and numbers is compiled to a comparison
in the Rust runtime/at the machine leve. Comparison for objects
is just a numeric comparison of the pointers. For Strings, that
will actually always work because the Strings are interned into
the `string` hashmap, so their memory location will be the same
for the same string always.

Comparison for objects can be put off to later if at all.

I also updated the values iterator for `Op` because the operations
weren't being converted back properly.