Add multiplication, division, change sig of interpret
A big change here: `interpret` now returns the VM whether or not
it encountered an error. That's not a big deal for interpreting
scripts, but for running in the repl, it means the repl no longer
crashes if something went wrong. I haven't though through the
implications of this for runtime exceptions, like if we stop in
the middle of executing bytecode, then we'll leave things on the
stack. Maybe we should just quit in the case of a runtime exception,
whereas parse or scan expressions are fine. Although they'll leave
garbage lying around. Anyways, I digress.

The multiplication and division are pretty straightforward, but
they led to the discover of a really fun bug, that I have isolated
to its own commit just after this one.