Don't run forever when encountering a single slash
I had a lot of fun writing the whitespace funcitons the first
time around because I learned about tags for loops in Rust
that let you break out of either an inner or outer loop conditionally,
which is super cool and probably something I could have used
in Java once or twice, if for no other reason then for clarity's sake.
However, I didn't actually write test cases for all of the corners
of the nested loops.

Namely, I didn't write a test for a simple division (`/`) operation.

Now today, when I added division, the program would hang forever.
My first though was maybe the division operation I wrote was wrong,
but debug print statements were never output when I added them to
the VM.

Then I thought maybe the problem was the compiler, because I
added debugs to the test runner and found the vm never executed,
but the compiler at least started. So I scanned around and couldn't
figure out where in the expression loop it was haninging, but I was
pretty sure it was in the expression loop because I'd get one print
output, then nothing!

So I kept looking, and eventually I realized there were parts of
the program where I could get two printed outputs. Aha! So it
wasn't the compiler, there must be another loop, maybe in
`advance`: I'd forgotten that it handles skipping whitespace,
including comments..

So this change is to the loop where comments are parsed, and when
we check whether a slash is a comment or is a division mark,
now we break out if it's a division mark, rather than *not
advancing the pointer forward* and starting over the loop!