Add test for variable declaration
This looks like a commit to add features to debug printing and fix
some mistakes I made the first time around, but really I just did
all that to figure out what the meaning of the opcodes that were
failing the test I added.

The tests basically passed as expected.