Define global variables operation
This is the runtime component of variable declaration, done in a
test-drivin way by adding a unit test that declares a variable.

This change just defines the variable in to a hash map (by string
name) of global variables. This works well with string interning
but does require us to store a second reference to every string.
We could store a hash map of the string locations instead. That
would be a good future improvement.