Implement global variable retrieval
This is exciting because now I am starting to get a real programming
language going. You can do math and store the result in vars that
can be used to do more math.

I'm definitely drifting from the book a bit in terms of function names,
which is definitely going to cause problems later. However, I feel
like this is rust-ish enough to be something I can be proud of rather
than just a stupid project. I'd enjoy extending this with some of the
book's optimizations.

This change also includes a bunch of lines to use arrays instead of
iterating over a Vec to translate between Op/Prec and u8, but that's
all just so that we don't waste a bunch of time on every loop over
the bytecode.