Compile block statements
In the LoxVM version of the lox interpreter, blocks from the
compiler's point of view are just a change to the tracking of the
scope (which we'll eventually use to resolve variables at
runtime), and then a bunch of expressions. To the VM, blocks
will not be tracked.

I honestly don't remember how local variables from a parent block
are referenced, so I'll have to revisit that when I implement it.

I've read back through the code and some bytecode generated by
the C version of the compiler, and things make a little sense,
but I'm still fuzzy.

The gist going forward of local variables is that they are
referenced by name at compile time, just like globals are
referenced by name at runtime by tracking the name in the
constants table (without interning because we don't know if
the item the name refers to exists in the global pool yet, so
we can't look for it).

Anyways, the locals at runtime are actually just referenced by
how far down the stack they go. And the block itself is not
really even a thing. The variable is put onto the stack, and
is referenced by its stack index, and there doesn't even need
to be a set local variable operation if the variable is declared,
and defined together, because it's already on the stack.

Finally, the `endScope` method takes care of cleaning up any
variables who have the current scope from the stack by just
emitting enough pop operations to clear them. Note, they don't
get popped until runtime, we're popping as many as we know will
be on the stack at runtime, because we've tracked all the local
variables in the compiler.