Set and Get local variables from the stack
This is done, as I said in my last commit message, by tracking the
local variables we use by how far down the stack they are and
at runtime we just start at the bottom of the stack (at the call
frame) and counting up to the desired local variable.

This is possible because the only instructions that leave values
on the stack or remove them are instructions that declare local
variables or that clear out of blocks or return from functions,
etc.

One more change in this blob is using function pointers when
debug printing, this is just to avoid so much duplicated code.