Get and set local variables at runtime
Most of the heavy lifting to get here was done in the last
commit, tracking where in the stack a variable will be by
using a shadow stack. This probably needs a few more changes:

[bug]

1. It needs to track how many frames we've used, or push a
   frame or pop it when we call a function, I don't remember
   how that works.
2. Global variables aren't set here, though that's pretty easy
3. We don't handle upvalues at all

I learned while doing this to definitely use the `insert`
method on Vec rather than trying to `mem::swap`. Doing swap
ended up just removing the value from the stack somehow, I'm not sure.