Very large change, stop using enums
The C version of this code uses C enums, and I am glad I
learned how to use `repr` to make Rust enums fit in a certain
memory size, but there's no need after all because converting
back and forth was just going to be expensive.

I ended up settling on an array because that would keep the order
correct, but it was still to tricky and prone to error.

This way of doing things guarantees that if I add a value in
the middle of the enum, it won't be covered. And if I add a
value on the end, hopefully I also add a unit test and notice
it isn't covered. That's a pretty big if, but once the
ops and precs are all settled, this shouldn't make any diff,
except probably to speed things up.