Update debug output to match new non-enum OpCodes
Since the opcodes are now just u8 values, I have to use the
`disassemble_byte` method I wrote for `u8` in order for them to
be output correctly. While I was working on that, I also broke
the logic to get the line number down into each of the printing
functions because that way it doesn't need to be a parameter of
each debugging operation.

Another benefit is that we can determine the line number from
the instruction pointer instead of `i` that's tracked in the
compiler, so the same logic (with some slight modifications)
can be re-used in the vm's debug trace execution.

This will be useful for debugging the runtime later on.