Keep the result of assignment on the stack
Because assignment is an expression, the result will be
popped by the `POP` operation emitted by
`expression_statement`.

This is done by just not popping the value off the stack.
I also think this would have been a bug that caused a
panic with the right test case so I'm glad I thought of
it :D