Concatenate strings at runtime and fix interning
The concatenation is relatively simple. We clone the first string,
add the second string, and then intern the result. Because this is
the first test that constructs a string at runtime, or compares a
string at all, it's the first time we've actually tested that the
interning code works.

The test in this commit exposed that bug, so the commit also includes
the fix, which is inserting the string into the strings map and
to memory for fast comparison. This also implies we can reverse
lookup the memory location of a string, though I doubt we'll
use that.