Synchronize after errors, re-use constants if possible
The synchronize method was bottomed with a `todo` macro, this
change just fills out that logic. The idea is to continue to a
symbol that indicates we're in a sane state to continue parsing,
and then parse the rest of the file to try and show as many helpful
errors as possible.

The constants change came about because I wanted to improve the error
handling of the constants table better, but when I did I realized
I could improve the creation of constants by checking if a chunk
already had that value. This is esepecially helpful for strings
because we'll already have interned variable names (for global
variables only, because we reference them by name at runtime), and
we don't need to track every reference to a global variable. This is
especially helpful in a large function where we reference many
global variables.