Compile if statements to jump instructions
This was one of the coolest moments in the book for me. It's where
the language starts to actually build out control flow primitives in
the bytecode so that we can do more interesting things with our
scripts.

The gist is that we compile an if statement by reading the condition,
then emitting a `JUMP_IF_FALSE` instruction. That instruction is
updated once we've read the whole block or statement to do if it
was true, because then we know where to jump if it was false.

Before the else block, we skip to the end so that only one of the
two is executed, and in the else block we just continue through
the end because we've already jumped once.