Create macro for running scripts in test
This is my first ever Rust macro (which is super exciting), and
I got the idea for it from this stack overflow[^1] which recommends
a similar approach for parameterized tests.

I think it may be possible to improve it further by deriving the test
name from the file name or vice versa, but this is fine for me for
now.

This also means I can really easily add new file tests in the
future and is the first chance for end-to-end tests of the
system. It also gives me some ideas on how I could improve other
tests, or how I could create macros to run the VM tests.

[1]: https://stackoverflow.com/questions/34662713/how-can-i-create-parameterized-tests-in-rust