Runtime JUMP_IF_FALSE, plus cleaning up run loop
The JUMP_IF_FALSE operation sets the next two bytes to exactly
how far we want to jump forward from the current instruction
pointer, so it's really easy to just read it and follow the
instruction. This is enough to implement a very simple if
statement without an else branch.

I also cleaned up the default branch of the main loop's op match
statement so that adding new operations will cause a compile
error, with the exception of adding operations at the end.