Fix patching, add tests around patching, runtime JUMP, function_mut, read_jump, and next_instruction
There's a lot going on in this change. I'll take it in the order
of the commit title.

### Fix patching and next_instruction

Patching was broken because I'd sort of guessed the right way to
record and adjust the patching locations when patching jumps, and
would only record the absolute minimum of locations. The fix was
instead fo add a helper that checks for the real value of the
instruction pointer during the execution of the most recent emitted
operation, and call that right before emitting the instructions to
be jumped from and to. This meant more local variables but the
end result is much more obvious when reading, I think.

### Runtime jump, add tests

This was the motivation behind the commit, and when testing the JUMP
instruction, I added tests that I realized would be useful in checking
the behavior of patching, which is why that is here too.

### Function_mut

I renamed the function borrowing method on VM to `function_mut` because
it borrows the function mutable, and added a non-mutable version,
which I needed to do to avoid borrow checker issues.

### Read_jump

The `read_jump` method just reads two bytes and turns them into a
`usize`, but it was broken because I'd just slipped up and used
and instead of or