For loops and add `emit_jump` helper
The `emit_jump` method just keeps me from having to type out
`emit_bytes(0, 0)` after every jump emitted. Since every
jump is going to be patched anyways, it might as well all
just share this method and make things easier.

The for loop was a little bit tricky; I had to look back at the C
code to remember the approach, but once I got the idea from the
frist half or so, the rest just kind of flowed out.

I don't have a compiler test for for loops because that would
totally suck to write