Compile a function declaration
This change ended up getting pretty hairy, then towrads the end
it turned out okay. I first tried to just take the parser and
scanner from the compiler directly, then tried using an enum to
borrow them all at once because taking them one at a time caused
errors due to multiple mutable borrows. But taking them all at
once also caused problems that couldn't be resolved re. lifetimes,
i.e. you can't borrow for part of the function if you use an
explicit lifetime of the whole struct.

There might have been a way to solve that second problem but it
would have been really hard to read. Instead I just require the
user to declare their parser and scanner separately. I could
create a macro that creates these for you, keeps them around
during compilation, and then only returns the VM. Or I could
even do that with a function, I guess.

In any case, this change isn't terribly exciting. It might as
well  just declare a variable to be nil when a function is
declared, you have to take it on faith that the value of that
function is some real and useful bytecode. The tests of that
(and I'm sure many changes to fix things that break) will come
next.