Compile a function call
The call operation also tracks the number of arguments the
function was called with. We may not know the arity of the
function at build time, in case it is a global function that gets
updated for example.

The call assumes that number of arguments is on the stack, and
that just before the arguments is the function itself, thus
we point the stack frame at runtime at the function value
when we call the function at runtime.