Runtime call function
The call operation is implemented by pushing a frame to the
call stack, and pointing that frame at the stack base, plus
keeping track of the next instruction pointer for when we
return.

This change looks like it's passing, but it doesn't actually work
because the `op::RETURN` instruction quits the program before
we can test the result, which is not the correct behavior.