[BUGFIX] Match function declaration to C version
I'm not sure what the complete set of issues with the previous
version are; I thought it had passed tests when I pushed it, but
I must have only run a subset. The key advance in this version is
re-using most of the Compiler constructor method in a new `spawn`
method that copies the fields from the parent.

This method then can be used to create the compiler, and instead of
creating a separate function, just emit the function location
at the end of compilation.