[BUGFIX] Working, remove final bugs from functions
There were a few doozies here. First, I added error logging for compilation
in VM tests. Second, I had called `function` instead of `fun` in the
VM tests. Third, I had pointed the frame to the fuction on the stack
instead of just above it. This doesn't work because we don't push the
script function to the stack when running to the program, so it's off
by one from the global case. Finally, all the call logic is already
refactored, so might as well re-use that!