Side step borrow checker with unsafe block
This is done inside a macro, so I feel like it's *super* unsafe
instead of just the regular unsafe, but I really want to use a
macro here (unlike the other macros because I want to match a
type that's passed in) and I can't emit the runtime error
without the unsafe block to de-ref the mut pointer, which I need
in order to alias the VM so that I can call mutable methods on
it.

I know that this is safe because the value that's borrowed is a
`None` so there's no way that I could conflict. I'm not sure if
this would cause a bunch of red flags to go off though.