[IN-PROGRESS] Added upvalues for closures
This commit is really big so I don't even think I can do it justice
here. The changes really should be documented in the code.

The basic outline is:

1. Modify macros to print rather than modify the error chains
2. Add upvalues and operation for creating closure

Though there's a handful of other things thrown in here as well.
It's been in progress for a few days, so I'm just going to create
and push this commit so I don't lose track of the work.