Fix global failure bug
Globals that have not been set are a legitimate use case of
the language. While the program should exit early in this case,
if we're in the REPL, we actually want to continue taking input
in case the user decides to set that global.

To do this, we just emit a runtime error rather than quitting
by panic.