Fix bug discovered by creatign crazy closure test
I created the test `crazy-closure.lox` in order to try to hit
as many edge cases as possible in one test with closures. I was
beating my head against it all night and discovered a few issues:

First, I noticed that I would compile the statement `vary x = 1;`,
which shouldn't work. At fist I thought it was the precedence rules,
so I went through those with a comb, but those changes are a no-op
because the only time precedence is used is when there's an infix
rule. The real problem was the test I was running was showing only
the `vm.error_chain`, so I didn't see the (already correct) error
messages. So I added all compile errors to the VM.

There actually was a bug with the scanning logic where `vary` was
parsed wrong, but it still was broken when I fixed that due to the
missing error message. Missing error message was just a test problem,
but the scanner is a real fix.

Second, I moved upvalues to the memory heap. This will mean they're
garbaged collected, so I try not to leave any pointers laying around
when I allocate anything, i.e. push things to the stack when they're
not being used yet but you still need to allocate something. I did
this so that I could track all the Upvalues together, and so that
the duplicates in `open_upvalues` and `closure.upvalues` would
actually point to the same thing.

I added `is_closed`, `is_after`, and `stack_index` helpers to the
upvalues. The stack index ones are useful for displaying and
sorting the upvalues. The `is_after` is necessary in order to
pop values off the open upvalue stack only while they're going
out of scope. `Is_closed` is used to skip over adding an upvalue
to the `open_upvalues` stack if its target has already been closed
previously.