Set upvalues at runtime
The compiler emitted set upvalue but I had not implemented it
yet. In order to test this functionality I made a familiar
counter closure which updates its value. This test ensures
that the closed variable is different for each closure, since
it's a different local each time counter is called, and that
it updates multiple times.

I could add another test like this to show that the same
variable in multiple closures is consistent.