Fix shared references inside of different closures
The key things I learned from this commit are best summed up in
the included comment, but the way I learned them was: I made this
new test expecting it to just work because upvalues are already
mostly pointers, but I was surprised that the counter kept different
references to the local variable 'x'.

I debugged by printing the upvalue before it was accessed, and I
noticed there were two different upvalue pointers in memory at
different locations, which means they probably were both closed separately,
and therefore I must have allocated two. So I debugged the open
upvalues on each call to allocate and noticed that the upvalue
existed, but wasn't shared. So this change just checks if an upvalue
ref to a local already exists, and if so then points to that.

There's no new indirection for the upvalues, they just point to
the same spot in memory.