Do not sort open upvalues
We save a ton of time by using a vector instead of a linked
list of upvalues, so we can iterate over them really fast.
We do have to do a memory lookup each time, but that's okay.
The real values here is that we can push a new upvalue really
quickly without any iteration, and we already had to iterate
through a bunch of upvalues before to close upvalues.

So it was already wost case O(n) removal and now it's O(n)
every time, but insertion is now O(1) always, while it was
previously O(n).

In addition to this change, I also removed all the places in
the code where we did something like `x.get(i).unwrap()` and
instead did a `x[i]` because in all of these cases we're
dealing with indices that were known at compile time.